backend/requirements.txt
Flask==1.1.4
Flask-Cors==3.0.9
Pillow==8.1.0
requests==2.25.1
python-dotenv==0.21.0
Werkzeug==1.0.1
textblob==0.15.3
markupsafe==2.0.1
Jinja2==2.11.3